--- dipy/tracking/local/__init__.py
@@ -1,15 +1,0 @@
-from .direction_getter import DirectionGetter
-from .localtracking import LocalTracking, ParticleFilteringTracking
-from .tissue_classifier import (ActTissueClassifier,
-                                BinaryTissueClassifier,
-                                CmcTissueClassifier,
-                                ConstrainedTissueClassifier,
-                                ThresholdTissueClassifier,
-                                TissueClassifier)
-
-from dipy.tracking import utils
-
-__all__ = ["ActTissueClassifier", "BinaryTissueClassifier",
-           "CmcTissueClassifier", "ConstrainedTissueClassifier",
-           "DirectionGetter", "LocalTracking", "ParticleFilteringTracking",
-           "ThresholdTissueClassifier", "TissueClassifier"]
