--- dipy/tracking/local/tests/__init__.py
@@ -1,1 +1,0 @@
-# Make tracking/local/tests a package
