+++ doc/examples/tracking_introduction_eudx.py
@@ -102,7 +102,7 @@ the ODF shows significant restricted diffusion by thresholding on
 the generalized fractional anisotropy (GFA).
 """
 
-from dipy.tracking.local import ThresholdTissueClassifier
+from dipy.tracking.tissue_classifier import ThresholdTissueClassifier
 
 classifier = ThresholdTissueClassifier(csa_peaks.gfa, .25)
 
@@ -153,7 +153,7 @@ used here and the default option when providing the output of peaks directly
 in LocalTracking.
 """
 
-from dipy.tracking.local import LocalTracking
+from dipy.tracking.localtracking import LocalTracking
 from dipy.tracking.streamline import Streamlines
 
 # Initialization of LocalTracking. The computation happens in the next step.
