+++ doc/examples/viz_roi_contour.py
@@ -13,9 +13,9 @@ from dipy.data import read_stanford_labels
 from dipy.reconst.shm import CsaOdfModel
 from dipy.data import default_sphere
 from dipy.direction import peaks_from_model
-from dipy.tracking.local import ThresholdTissueClassifier
+from dipy.tracking.tissue_classifier import ThresholdTissueClassifier
 from dipy.tracking import utils
-from dipy.tracking.local import LocalTracking
+from dipy.tracking.localtracking import LocalTracking
 from dipy.tracking.streamline import Streamlines
 from dipy.viz import actor, window, colormap as cmap
