+++ setup.py
@@ -77,9 +77,9 @@ for modulename, other_sources, language in (
         ('dipy.reconst.quick_squash', [], 'c'),
         ('dipy.tracking.distances', [], 'c'),
         ('dipy.tracking.streamlinespeed', [], 'c'),
-        ('dipy.tracking.local.localtrack', [], 'c'),
-        ('dipy.tracking.local.direction_getter', [], 'c'),
-        ('dipy.tracking.local.tissue_classifier', [], 'c'),
+        ('dipy.tracking.localtrack', [], 'c'),
+        ('dipy.tracking.direction_getter', [], 'c'),
+        ('dipy.tracking.tissue_classifier', [], 'c'),
         ('dipy.tracking.vox2track', [], 'c'),
         ('dipy.tracking.propspeed', [], 'c'),
         ('dipy.tracking.fbcmeasures', [], 'c'),
@@ -192,8 +192,6 @@ def main(**extra_args):
                     'dipy.direction',
                     'dipy.direction.tests',
                     'dipy.tracking',
-                    'dipy.tracking.local',
-                    'dipy.tracking.local.tests',
                     'dipy.tracking.tests',
                     'dipy.tracking.benchmarks',
                     'dipy.reconst',
